hook everything up
