+++ lib/llm/src/mocker/protocols.rs
@@ -97,6 +97,14 @@ pub struct MockEngineArgs {
     /// Optional startup time in seconds to simulate engine initialization delay
     #[builder(default = "None")]
     pub startup_time: Option<f64>,
+
+    /// Whether to publish KV events (disabled for decode workers)
+    #[builder(default = "true")]
+    pub publish_kv_events: bool,
+
+    /// Whether this is a prefill worker (limits max_tokens to 1, returns dummy disaggregated_params)
+    #[builder(default = "false")]
+    pub is_prefill: bool,
 }
 
 impl Default for MockEngineArgs {
@@ -132,6 +140,8 @@ impl MockEngineArgs {
             "speedup_ratio",
             "dp_size",
             "startup_time",
+            "publish_kv_events",
+            "is_prefill",
         ]
         .iter()
         .cloned()
@@ -211,6 +221,18 @@ impl MockEngineArgs {
             && let Some(num) = value.as_f64()
         {
             builder = builder.startup_time(Some(num));
+        }
+
+        if let Some(value) = extra_args.get("publish_kv_events")
+            && let Some(enabled) = value.as_bool()
+        {
+            builder = builder.publish_kv_events(enabled);
+        }
+
+        if let Some(value) = extra_args.get("is_prefill")
+            && let Some(enabled) = value.as_bool()
+        {
+            builder = builder.is_prefill(enabled);
         }
 
         // Build the MockEngineArgs with either defaults or overridden values
